+++ module/data/actor/agent.js
@@ -1,5 +1,6 @@
 import HumanSkillsActorData from "./base/human-skills.js";
 import CharacterData from "./base/character.js";
+import DGHTMLField from "../fields/html-content-field.js";
 
 const { SchemaField, NumberField, StringField, BooleanField, HTMLField } =
   foundry.data.fields;
@@ -29,11 +30,7 @@ export default class AgentData extends CharacterData {
         }),
       }),
       physical: new SchemaField({
-        description: new HTMLField({
-          initial: "",
-          blank: true,
-          textSearch: true,
-        }),
+        description: DGHTMLField(),
         wounds: new StringField({ initial: "" }),
         firstAidAttempted: new BooleanField({ initial: false }),
         exhausted: new BooleanField({ initial: false }),