+++ module/data/actor/npc.js
@@ -1,5 +1,6 @@
 import HumanSkillsActorData from "./base/human-skills.js";
 import CharacterData from "./base/character.js";
+import DGHTMLField from "../fields/html-content-field.js";
 
 const { SchemaField, NumberField, StringField, BooleanField, HTMLField } =
   foundry.data.fields;
@@ -19,7 +20,7 @@ export default class NPCData extends CharacterData {
       biography: new SchemaField({
         profession: new StringField({ initial: "" }),
       }),
-      notes: new HTMLField({ initial: "", blank: true, textSearch: true }),
+      notes: DGHTMLField(),
       shortDescription: new StringField({ initial: "" }),
       showUntrainedSkills: new BooleanField({ initial: true }),
     };