+++ module/data/actor/vehicle.js
@@ -1,4 +1,5 @@
 import SheetSettingsActorData from "./base/sheet-settings.js";
+import DGHTMLField from "../fields/html-content-field.js";
 
 const { SchemaField, NumberField, StringField, ArrayField, HTMLField } =
   foundry.data.fields;
@@ -8,11 +9,7 @@ export default class VehicleData extends foundry.abstract.TypeDataModel {
     return {
       ...SheetSettingsActorData.defineSchema(),
       name: new StringField({ initial: "" }),
-      description: new HTMLField({
-        initial: "",
-        blank: true,
-        textSearch: true,
-      }),
+      description: DGHTMLField(),
       shortDescription: new StringField({ initial: "" }),
       health: new SchemaField({
         value: new NumberField({ initial: 10 }),