+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 ### **Bug Fixes:**
 - [#392](https://github.com/deltagreen-foundryvtt/delta-green-foundry-vtt-system/issues/392) - Fixed sheet rolls
+- [#394](https://github.com/deltagreen-foundryvtt/delta-green-foundry-vtt-system/issues/394) - Unified actor and item description fields on schema with shared ProseMirror preparation
 
 ## Version 1.6.6 - 2026-04-15